Some mods and addons for better highligths, search, reading and customization

+++ Source/Models/Highlight.swift
@@ -16,11 +16,14 @@ class Highlight: NSManagedObject {
     @NSManaged var content: String
     @NSManaged var contentPost: String
     @NSManaged var contentPre: String
+    @NSManaged var startPos : NSNumber
+    @NSManaged var endPos : NSNumber
     @NSManaged var date: NSDate
     @NSManaged var highlightId: String
     @NSManaged var page: NSNumber
     @NSManaged var type: NSNumber
-
+    @NSManaged var notes: String
+    @NSManaged var childCount: NSNumber
 }
 
 public typealias Completion = (error: NSError?) -> ()
@@ -28,7 +31,7 @@ let coreDataManager = CoreDataManager()
 
 extension Highlight {
     
-    static func persistHighlight(object: FRHighlight, completion: Completion?) {
+    static func persistHighlight(object: FRHighlight, completion: Completion?, note : String) {
         var highlight: Highlight?
         
         do {
@@ -46,6 +49,10 @@ extension Highlight {
             highlight!.contentPost = object.contentPost
             highlight!.date = object.date
             highlight!.type = object.type.hashValue
+            highlight!.notes = note
+            highlight!.startPos = object.startPos
+            highlight!.endPos = object.endPos
+            highlight!.childCount = object.childCount
         } else {
             highlight = NSEntityDescription.insertNewObjectForEntityForName("Highlight", inManagedObjectContext: coreDataManager.managedObjectContext) as? Highlight
             coreDataManager.saveContext()
@@ -58,8 +65,11 @@ extension Highlight {
             highlight!.highlightId = object.id
             highlight!.page = object.page
             highlight!.type = object.type.hashValue
+            highlight!.notes = note
+            highlight!.startPos = object.startPos
+            highlight!.endPos = object.endPos
+            highlight!.childCount = object.childCount
         }
-
         // Save
         do {
             try coreDataManager.managedObjectContext.save()
@@ -72,6 +82,32 @@ extension Highlight {
             }
         }
     }
+
+    static func persistHighlight(){
+    
+        // Save
+        do {
+            try coreDataManager.managedObjectContext.save()
+
+        } catch let error as NSError {
+            print(error)
+        }
+    
+    }
+    
+    static func getHighlightById(highId : String) -> Highlight? {
+        var highlight: Highlight?
+        do {
+            let fetchRequest = NSFetchRequest(entityName: "Highlight")
+            fetchRequest.predicate = NSPredicate(format:"highlightId = %@", highId)
+            
+            highlight = try coreDataManager.managedObjectContext.executeFetchRequest(fetchRequest).last as? Highlight
+            return highlight
+        }catch let error as NSError {
+            print("Error on remove highlight: \(error)")
+        }
+        return nil
+    }
     
     static func removeById(highlightId: String) {
         var highlight: Highlight?
@@ -81,10 +117,13 @@ extension Highlight {
             fetchRequest.predicate = NSPredicate(format:"highlightId = %@", highlightId)
             
             highlight = try coreDataManager.managedObjectContext.executeFetchRequest(fetchRequest).last as? Highlight
-            coreDataManager.managedObjectContext.deleteObject(highlight!)
-            coreDataManager.saveContext()
+            
+            if highlight != nil {
+                coreDataManager.managedObjectContext.deleteObject(highlight!)
+                coreDataManager.saveContext()
+            }
         } catch let error as NSError {
-            print("Error on remove highlight: \(error)")
+            print("------>Error on remove highlight: \(error)")
         }
     }
     
@@ -96,6 +135,12 @@ extension Highlight {
             fetchRequest.predicate = NSPredicate(format:"highlightId = %@", highlightId)
             
             highlight = try coreDataManager.managedObjectContext.executeFetchRequest(fetchRequest).last as? Highlight
+            
+            let replaced = "class=\"\(HighlightStyle.classForStyle(highlight!.type.hashValue))\""
+            let replace = "class=\"\(HighlightStyle.classForStyle(type.hashValue))\""
+            var content = highlight!.content
+            content = content.stringByReplacingOccurrencesOfString(replaced, withString: replace)
+            highlight?.content = content
             highlight?.type = type.hashValue
             coreDataManager.saveContext()
         } catch let error as NSError {
